+++ pygmo/expose_islands.cpp
@@ -8,6 +8,7 @@
 
 #include <pybind11/pybind11.h>
 
+#include <pagmo/config.hpp>
 #include <pagmo/island.hpp>
 #include <pagmo/islands/thread_island.hpp>
 
@@ -53,7 +54,11 @@ void expose_islands(py::module &m, py::class_<pagmo::island> &isl, py::module &i
     test_isl.def("set_n", &detail::test_island::set_n);
 
     // Thread island.
-    expose_island<pagmo::thread_island>(m, isl, isl_module, "thread_island", thread_island_docstring().c_str());
+    auto thread_island_
+        = expose_island<pagmo::thread_island>(m, isl, isl_module, "thread_island", thread_island_docstring().c_str());
+#if PAGMO_VERSION_MAJOR > 2 || (PAGMO_VERSION_MAJOR == 2 && PAGMO_VERSION_MINOR >= 16)
+    thread_island_.def(py::init<bool>(), py::arg("use_pool") = true);
+#endif
 }
 
 } // namespace pygmo